Reflections chapter done good job

+++ src/color_calculation.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   color_calculation.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: udumas <udumas@student.42.fr>              +#+  +:+       +#+        */
+/*   By: vda-conc <vda-conc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/21 16:30:27 by udumas            #+#    #+#             */
-/*   Updated: 2024/05/22 20:19:07 by udumas           ###   ########.fr       */
+/*   Updated: 2024/05/29 19:01:07 by vda-conc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,10 +32,10 @@ t_color	ft_dif_color(t_color c1, t_color c2)
     return (new);
 }
 
-t_color ft_mult_color(t_color c, int scalar)
+t_color ft_mult_color(t_color c, float scalar)
 {
     t_color new;
-    
+
     new.r = c.r * scalar;
     new.g = c.g * scalar;
     new.b = c.b * scalar;
